Add new e2e tests and adjust playwright config
New e2e tests have been added for the snack-bar-demo and edit-svg-icon-demo. These tests ensure the visibility of success and error snack bars, and the edit icon button respectively.

Adjustments were made to the playwright.config.ts file to improve test execution. Commented out the global timeout and reduced the expect timeout to 10 seconds to fail fast if an assertion is not met within the timeframe. The number of retries was set to zero.

Changes were made in the home.spec.ts file as well. The css locator has been fixed from 'css=btn-close' to '.btn-close' for tests checking the presence of 'Success Alert' and 'Error Alert'. This was done to correctly locate the close button for these alerts.