Update AutocompleteComponent to use ControlValueAccessor
The update involves substantial refactor on AutocompleteComponent to properly utilize ControlValueAccessor feature from Angular forms. This ensures a seamless data flow between the formControl and the component while sustaining accessibility and consistency. This change enhances the component's reusability by allowing it to work more readily with reactive forms.

An associated HTML template was updated to reflect these changes. The `inputFormGroup` was removed from the formGroup and the value check was updated from form group specific to a more general check.

Also, the `NG_VALUE_ACCESSOR` was introduced in the providers to inject the instance of the AutocompleteComponent as a ControlValueAccessor. This is necessary to allow the AutocompleteComponent to interact with the ngModel directive.

Enhancements include writing new values to the component, registering an onChange callback to update the form model when the control's value changes, registering a onTouched callback to update the form model on blur and also to handle the disable state of the component.