chore: Add .browserslistrc file to specify supported browsers
A .browserslistrc file has been created to specify the list of supported browsers for the project. This file configures the build system to adjust CSS and JavaScript output in a way that they can support specified browsers.

The supporters are 10 different versions of Chromium-based browsers, Firefox, Safari, iOS, and Microsoft Edge. This addition is crucial to ensure that our application remains compatible and provide a uniform experience across the specified browsers.