chore: Simplified Github Build workflow and corrected README badge link
This commit contains two changes primarily. In the Github workflow file build.yml, the 'Build Project' name is simplified to 'Build'. The change is made to clean up and streamline the indicator names in our CI/CD pipeline. Secondly, in the README.md, the image URL for the CI badge is updated to point towards the new 'Build' workflow. The link previously pointed to a non-existent 'cicd.yml' file and this correction ensures users see the correct build status when visiting the project on Github.