enhance(alert): Add supporting custom styles to ngxsmart alerts
Added a new @Input() property 'class' to alert.component in ngxsmart to allow addition of custom classes to the alert component. This would enable developers to apply more specific styles on the component if needed. The usage is updated in the demo HTML file. Also, minor adjustments are made in the alert-demo.component like changing the initially set dismissOnTimeout property to false, probably done for testing the new feature.